updated layout for bokeh script sans django
